moved all thestar scrap file into the_star folder
