+++ scraping/the_star/Firebase/__init__.py
@@ -1,0 +1,4 @@
+"""
+    Copyright 2020 Edmund Hee, All Right Reserved
+    Email: edmund.hee05@gmail.com
+"""
