+++ scraping/the_star/Firebase/firebase_model.py
