+++ scraping/the_star/requirements.txt
